refactor solution
