+++ frontend/app/pages/SessionPage.tsx
@@ -1,8 +1,7 @@
-import Combobox from '@/components/combobox'
 import ConfirmationButton from '@/components/ConfirmationButton'
-import { PageSubheader } from '@/components/PageHeader'
 import SessionCard from '@/components/session/SessionCard'
 import SessionForm from '@/components/session/SessionForm'
+import SessionSignupPanel from '@/components/session/SessionSignupPanel'
 import TournamentList from '@/components/tournament/TournamentList'
 import TrackLeaderboard from '@/components/track/TrackLeaderboard'
 import {
@@ -23,341 +22,16 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog'
-import { Empty } from '@/components/ui/empty'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
 import { Spinner } from '@/components/ui/spinner'
-import UserRow from '@/components/user/UserRow'
 import { useAuth } from '@/contexts/AuthContext'
 import { useConnection } from '@/contexts/ConnectionContext'
 import { useData } from '@/contexts/DataContext'
 import loc from '@/lib/locales'
-import { userToLookupItem } from '@/lib/lookup-utils'
-import type { SessionWithSignups } from '@common/models/session'
-import type { UserInfo } from '@common/models/user'
-import { isUpcoming } from '@common/utils/date'
-import {
-  CircleCheck,
-  CircleQuestionMark,
-  CircleX,
-  PencilIcon,
-  Trash2,
-  type LucideIcon,
-} from 'lucide-react'
-import { useMemo, useState, type ComponentProps } from 'react'
+import { PencilIcon, Trash2 } from 'lucide-react'
+import { useState } from 'react'
 import { useParams } from 'react-router'
 import { toast } from 'sonner'
-import { twMerge } from 'tailwind-merge'
-import type { SessionResponse } from '../../../backend/database/schema'
 import { SubscribeButton } from './SessionsPage'
-
-const RESPONSE_OPTIONS: { response: SessionResponse; Icon: LucideIcon }[] = [
-  { response: 'yes', Icon: CircleCheck },
-  { response: 'maybe', Icon: CircleQuestionMark },
-  { response: 'no', Icon: CircleX },
-]
-
-function getUserSortName(user: UserInfo) {
-  return user.shortName ?? `${user.firstName} ${user.lastName ?? ''}`
-}
-
-function Signup({
-  session,
-  disabled,
-  className,
-  ...rest
-}: Readonly<
-  { session: SessionWithSignups; disabled?: boolean } & ComponentProps<'div'>
->) {
-  const { socket } = useConnection()
-  const { loggedInUser, isLoggedIn } = useAuth()
-  const { users, isLoadingData } = useData()
-
-  const myResponse = session.signups.find(
-    s => s.user.id === loggedInUser?.id
-  )?.response
-
-  const canManageSignups = isLoggedIn && loggedInUser.role !== 'user'
-  const [addDialogOpen, setAddDialogOpen] = useState(false)
-  const [selectedUserIds, setSelectedUserIds] = useState<string[]>([])
-  const [selectedResponse, setSelectedResponse] =
-    useState<SessionResponse>('yes')
-
-  const sortedSignups = useMemo(
-    () =>
-      session.signups.toSorted((a, b) =>
-        getUserSortName(a.user).localeCompare(getUserSortName(b.user))
-      ),
-    [session.signups]
-  )
-  const signedUpUserIds = useMemo(
-    () => new Set(session.signups.map(s => s.user.id)),
-    [session.signups]
-  )
-  const availableUsers = useMemo(
-    () =>
-      users
-        ?.filter(user => !signedUpUserIds.has(user.id))
-        .toSorted((a, b) =>
-          getUserSortName(a).localeCompare(getUserSortName(b))
-        ) ?? [],
-    [users, signedUpUserIds]
-  )
-  const selectedUsers = useMemo(
-    () => availableUsers.filter(user => selectedUserIds.includes(user.id)),
-    [availableUsers, selectedUserIds]
-  )
-  const selectableUsers = useMemo(
-    () => availableUsers.filter(user => !selectedUserIds.includes(user.id)),
-    [availableUsers, selectedUserIds]
-  )
-
-  if (isLoadingData)
-    return (
-      <div className='flex h-32 w-full items-center-safe justify-center-safe rounded-sm border border-border'>
-        <Spinner className='size-6' />
-      </div>
-    )
-
-  async function updateSignup(response: SessionResponse, user: UserInfo) {
-    const res = await socket.emitWithAck('rsvp_session', {
-      type: 'RsvpSessionRequest',
-      session: session.id,
-      user: user.id,
-      response,
-    })
-
-    if (!res.success) throw new Error(res.message)
-  }
-
-  function handleRsvp(response: SessionResponse, user?: UserInfo) {
-    if (!isLoggedIn) return
-    toast.promise(updateSignup(response, user ?? loggedInUser), {
-      loading: loc.no.session.rsvp.response.loading,
-      success: loc.no.session.rsvp.response.success(response),
-      error: loc.no.session.rsvp.response.error,
-    })
-  }
-
-  function handleSelectUser(user: UserInfo | null | undefined) {
-    if (!user) return
-    setSelectedUserIds(current =>
-      current.includes(user.id) ? current : [...current, user.id]
-    )
-  }
-
-  function removeSelectedUser(userId: string) {
-    setSelectedUserIds(current =>
-      current.filter(selectedUserId => selectedUserId !== userId)
-    )
-  }
-
-  function handleAddParticipants() {
-    if (selectedUsers.length === 0) return
-
-    toast.promise(
-      Promise.all(
-        selectedUsers.map(user => updateSignup(selectedResponse, user))
-      ).then(() => {
-        setSelectedUserIds([])
-        setAddDialogOpen(false)
-      }),
-      loc.no.session.rsvp.manage.addRequest(selectedUsers.length)
-    )
-  }
-
-  return (
-    <div className={twMerge('flex flex-col gap-4', className)} {...rest}>
-      <div className='flex justify-between'>
-        <h3 className='px-2 pt-2'>
-          {isUpcoming(session)
-            ? loc.no.session.attendance
-            : loc.no.session.attendees}
-        </h3>
-
-        <div>
-          {(isUpcoming(session) || canManageSignups) &&
-            isLoggedIn &&
-            myResponse && (
-              <Select
-                disabled={disabled}
-                value={myResponse}
-                onValueChange={handleRsvp}>
-                <SelectTrigger className='w-[160px]'>
-                  <SelectValue placeholder={loc.no.session.rsvp.change} />
-                </SelectTrigger>
-                <SelectContent>
-                  {RESPONSE_OPTIONS.map(({ response, Icon }) => (
-                    <SelectItem key={response} value={response}>
-                      <Icon className='size-4' />
-                      {loc.no.session.rsvp.responses[response]}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            )}
-        </div>
-      </div>
-
-      <div
-        className='flex items-center justify-center gap-2'
-        hidden={!isUpcoming(session) || !isLoggedIn || !!myResponse}>
-        {RESPONSE_OPTIONS.map(({ response, Icon }) => (
-          <Button
-            key={response}
-            size='sm'
-            onClick={() => handleRsvp(response)}
-            disabled={disabled}>
-            <Icon className='size-4' />
-            {loc.no.session.rsvp.responses[response]}
-          </Button>
-        ))}
-      </div>
-
-      {canManageSignups && (
-        <Dialog open={addDialogOpen} onOpenChange={setAddDialogOpen}>
-          <DialogTrigger asChild>
-            <Button type='button' variant='outline' disabled={disabled}>
-              <CircleCheck className='size-4' />
-              {loc.no.session.rsvp.manage.add}
-            </Button>
-          </DialogTrigger>
-          <DialogContent>
-            <DialogHeader>
-              <DialogTitle>{loc.no.session.rsvp.manage.title}</DialogTitle>
-            </DialogHeader>
-            <div className='flex flex-col gap-4'>
-              <Combobox
-                placeholder={loc.no.session.rsvp.manage.userPlaceholder}
-                emptyLabel={loc.no.common.noItems ?? undefined}
-                items={selectableUsers.map(userToLookupItem)}
-                selected={undefined}
-                setSelected={handleSelectUser}
-                CustomRow={props => <UserRow {...props} hideLink hideRanking />}
-              />
-
-              {availableUsers.length === 0 && (
-                <Empty className='border border-input text-sm text-muted-foreground'>
-                  {loc.no.common.noItems}
-                </Empty>
-              )}
-
-              {selectedUsers.length > 0 && (
-                <div className='max-h-80 overflow-y-auto rounded-sm bg-background-secondary'>
-                  {selectedUsers.map(user => (
-                    <div
-                      key={user.id}
-                      className='flex items-center gap-2 border-b border-border/60 p-2 last:border-b-0'>
-                      <UserRow
-                        item={user}
-                        hideLink
-                        hideRanking
-                        className='flex-1 py-2'
-                      />
-                      <Button
-                        type='button'
-                        size='icon'
-                        variant='ghost'
-                        onClick={() => removeSelectedUser(user.id)}>
-                        <CircleX className='size-4' />
-                      </Button>
-                    </div>
-                  ))}
-                </div>
-              )}
-              <Select
-                value={selectedResponse}
-                onValueChange={value =>
-                  setSelectedResponse(value as SessionResponse)
-                }>
-                <SelectTrigger className='w-full'>
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  {RESPONSE_OPTIONS.map(({ response, Icon }) => (
-                    <SelectItem key={response} value={response}>
-                      <Icon className='size-4' />
-                      {loc.no.session.rsvp.responses[response]}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-            <DialogFooter>
-              <DialogClose asChild>
-                <Button variant='outline'>{loc.no.common.cancel}</Button>
-              </DialogClose>
-              <Button
-                type='button'
-                onClick={handleAddParticipants}
-                disabled={selectedUserIds.length === 0}>
-                {loc.no.session.rsvp.manage.addSelected}
-              </Button>
-            </DialogFooter>
-          </DialogContent>
-        </Dialog>
-      )}
-
-      {sortedSignups.length === 0 && (
-        <Empty className='border border-input text-sm text-muted-foreground'>
-          {loc.no.common.noItems}
-        </Empty>
-      )}
-
-      {RESPONSE_OPTIONS.map(({ response }) => {
-        const responseSignups = sortedSignups.filter(
-          s => s.response === response
-        )
-        if (responseSignups.length === 0) return undefined
-        return (
-          <div key={response} className='flex flex-col'>
-            <PageSubheader
-              title={loc.no.session.rsvp.responses[response]}
-              description={responseSignups.length.toString()}
-            />
-            <div className='rounded-sm bg-background-secondary'>
-              {responseSignups.map(({ user }) => {
-                return (
-                  <UserRow
-                    key={user.id}
-                    item={user}
-                    className='w-full py-1 first:pt-2 last:pb-2'
-                    hideRanking>
-                    {canManageSignups && (
-                      <Select
-                        value={response}
-                        onValueChange={value =>
-                          handleRsvp(value as SessionResponse, user)
-                        }
-                        disabled={disabled}>
-                        <SelectTrigger>
-                          <SelectValue />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {RESPONSE_OPTIONS.map(({ response, Icon }) => (
-                            <SelectItem key={response} value={response}>
-                              <Icon className='size-4' />
-                              {loc.no.session.rsvp.responses[response]}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    )}
-                  </UserRow>
-                )
-              })}
-            </div>
-          </div>
-        )
-      })}
-    </div>
-  )
-}
 
 export default function SessionPage() {
   const { id } = useParams()
@@ -472,7 +146,7 @@ export default function SessionPage() {
         )}
       </div>
 
-      <Signup
+      <SessionSignupPanel
         className='rounded-sm border bg-background p-2'
         disabled={isCancelled}
         session={session}
